Added ShellRunnerFactory and IShellRunner is used instead ShellRunner.
